ai-service/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pandas==2.3.3
numpy==2.2.6
scikit-learn==1.7.2
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.12.2
